Fix build and add CI (#1)
- use math defines
- correct some signed/unsigned comparaison warnings on Windows
- update [pa.delay5~] object
- correct ‘for’ loop initial declarations that are only allowed in C99 mode
- Add CI badges

+++ source/projects/pa.delay3_tilde/pa.delay3_tilde.c
@@ -19,8 +19,8 @@ typedef struct _pa_delay3_tilde
     float*      m_buffer;
     int         m_buffersize;
     
-    size_t      m_writer_playhead;
-    size_t      m_reader_playhead;
+    int         m_writer_playhead;
+    int         m_reader_playhead;
     
     t_outlet*   m_out;
